Código de inteface movido para paco 'ui'
close #50
