+++ crawlers/reddit_crawler.py
@@ -1,5 +1,4 @@
 import requests
-import click
 
 BASE_URL = "https://www.reddit.com"
 
@@ -70,15 +69,3 @@ def print_subreddits(threads):
 
 def filter_by_votes(threads, min_votes=1):
     return [thread for thread in threads if thread['upvotes'] > min_votes]
-
-
-@click.command()
-@click.option('--subreddits', '-s', required=True, help="Threads que deseja buscar separadas por ';'")
-@click.option('--min-votes', '-m', default=5000, help="Número mínimo de votos")
-def main(subreddits, min_votes):
-    for subreddit in subreddits.split(';'):
-        print_subreddits(filter_by_votes(get_threads(subreddit), min_votes=min_votes))
-
-
-if __name__ == '__main__':
-    main()
